fix: ui 레이어에서 구현하도록 변경
