--- Android/app/src/main/java/com/team16/airbnb/data/repository/CalendarRepositoryImpl.kt
@@ -1,17 +1,0 @@
-package com.team16.airbnb.data.repository
-
-import com.team16.airbnb.data.datasource.CalendarDataSourceImpl
-import com.team16.airbnb.data.model.CalendarData
-import com.team16.airbnb.data.model.DayInfo
-import kotlinx.coroutines.flow.Flow
-import javax.inject.Inject
-
-class CalendarRepositoryImpl @Inject constructor(
-    private val calendarDataSource: CalendarDataSourceImpl
-): CalendarRepository {
-
-    override fun getCalendar() = calendarDataSource.getCalendar()
-
-    override fun setPickDate(startDate: DayInfo, endDate: DayInfo) = calendarDataSource.setPickDate(startDate, endDate)
-
-}
