+++ Android/app/src/main/java/com/team16/airbnb/di/DataSourceModule.kt
@@ -1,21 +1,12 @@
 package com.team16.airbnb.di
 
-import com.team16.airbnb.data.datasource.CalendarDataSource
-import com.team16.airbnb.data.datasource.CalendarDataSourceImpl
-import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
 abstract class DataSourceModule {
 
-    @Singleton
-    @Binds
-    abstract fun bindCalendarDataSource(
-        calendarDataSourceImpl: CalendarDataSourceImpl
-    ) : CalendarDataSource
 
 }
