+++ Android/app/src/main/java/com/team16/airbnb/ui/calendar/CalendarViewModel.kt
@@ -2,20 +2,19 @@ package com.team16.airbnb.ui.calendar
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.team16.airbnb.common.DateChoiceListener
 import com.team16.airbnb.data.model.CalendarData
 import com.team16.airbnb.data.model.DayInfo
-import com.team16.airbnb.data.repository.CalendarRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.launch
+import java.util.*
 import javax.inject.Inject
 
 @HiltViewModel
-class CalendarViewModel @Inject constructor(
-    private val calendarRepository: CalendarRepository
-) : ViewModel() {
+class CalendarViewModel @Inject constructor() : ViewModel() {
 
     private val _calendar = MutableStateFlow<List<CalendarData>>(emptyList())
     val calendar: StateFlow<List<CalendarData>> = _calendar
@@ -30,15 +29,88 @@ class CalendarViewModel @Inject constructor(
 
     private var endDate = DayInfo(day = "", month = "")
 
+    private var todayMonth: Int = -1
+    private var today: Int = -1
 
+    private var id: Int = 0
 
     init {
         viewModelScope.launch {
-            calendarRepository.getCalendar().collect {
+            getCalendar().collect {
                 _calendar.value = it
             }
         }
     }
+
+    private fun getCalendar(): Flow<List<CalendarData>> = flow {
+        val cal = Calendar.getInstance()
+        var month = cal.get(Calendar.MONTH) + 1
+        var year = cal.get(Calendar.YEAR)
+
+        todayMonth = month
+        today = cal.get(Calendar.DATE)
+
+        val list = mutableListOf<CalendarData>()
+
+        for (i in 0 until 12) {
+            if (month > 12) {
+                month = 1
+                year++
+            }
+
+            list.add(
+                CalendarData(
+                    "${year}년 ${month}월",
+                    setDays(year, month)
+                )
+            )
+            month++
+        }
+        emit(list)
+    }
+
+    private fun setDays(year: Int, month: Int): Map<Int, DayInfo> {
+        val cal = Calendar.getInstance()
+        cal.set(year, month - 1, 1)
+        var start = cal.get(Calendar.DAY_OF_WEEK)
+        val days = mutableMapOf<Int, DayInfo>()
+
+        for (i in 1 until start) {
+            days[id] = DayInfo(id," ", " ", isPossible = false)
+            id++
+        }
+
+        for (i in 1..cal.getActualMaximum(Calendar.DATE)) {
+            when (month == todayMonth) {
+                true -> {
+                    days[id] = DayInfo(
+                        id = id,
+                        day = "$i",
+                        month = "$month",
+                        isPossible = setIsPossible(today, i)
+                    )
+                    id++
+                }
+
+                false -> {
+                    days[id] =
+                        DayInfo(
+                            id = id,
+                            day = "$i",
+                            month = "$month",
+                            isPossible = true
+                        )
+                    id++
+                }
+            }
+
+            start++
+        }
+
+        return days
+    }
+
+    private fun setIsPossible(today: Int, day: Int) = today <= day
 
     fun setDate(dayInfo: DayInfo) {
 
@@ -105,6 +177,7 @@ class CalendarViewModel @Inject constructor(
     private fun setPickDate() {
 
         val list = getCopyList()
+
         list.forEach { data ->
 
             data.days[startDate.id]?.apply {
@@ -132,6 +205,7 @@ class CalendarViewModel @Inject constructor(
         val list = getCopyList()
 
         list.forEach { data ->
+
             for (i in startDate.id..endDate.id) {
                 data.days[i]?.apply {
                     isChoice = false
@@ -142,10 +216,12 @@ class CalendarViewModel @Inject constructor(
             }
 
         }
+
         startDate = DayInfo(day = "", month = "")
         endDate = DayInfo(day = "", month = "")
         _pickState.value = false
         _calendar.value = list
+        _label.value = ""
     }
 
     private fun getCopyList(): List<CalendarData> {
