+++ Android/app/src/main/res/layout/item_calendar.xml
@@ -34,7 +34,7 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintWidth_percent="0.89"
-            tools:listitem="@layout/item_days"/>
+            tools:listitem="@layout/item_days" />
 
 
     </androidx.constraintlayout.widget.ConstraintLayout>
